requirements.txt
potassium>=0.0.9
torch
transformers
Pillow
numpy